Align sample filtering requirements with previous augur builds
Adds a new filter rule that applies the augur filter command to omit sequences
shorter than a minimum length, outliers, and samples with missing geographic
metadata prior to strain selection, in an attempt to match the filtering
requirements used by previous augur builds [1]. This commit also updates
`select_strains.py` to accept one or more FASTA files and to only consider
metadata for samples whose sequences made it through the aforementioned
filters. After strains are selected, we still need a rule to extract the segment
sequences corresponding to each strain. This rule has been renamed from "filter"
to "extract" and its body has been replaced with a call to a standalone script
that performs the extraction using the same logic as the original Python code.

On a related note, this commit also imposes an upper date limit on included
strains in `select_strains.py` to prevent reference strains from the future from
being included in a build. This modification has no effect on standard "live"
builds, but it ensures consistent output for custom builds that request a
specific time interval from the past (e.g., 2005-2015) that should not be
assigned any sequences after the last date in the requested time interval.

Resolves #6.

[1] https://github.com/nextstrain/augur/blob/3.1.0/builds/flu/flu.prepare.py#L102-L112