Ensure no future sequences are included in builds
Fixes a bug in the logic associated with filtering strains by date which was
causing strains from outside requested time intervals to be included in the
final output. This bug was due to an off-by-one error in parsing of months where
metadata month values were zero-indexed and being compared to one-indexed months
in datetime objects.

This commit addresses this error by filtering strains by numeric float dates at
the beginning of the flu subsampling function. This approach allows users to
specify time intervals to the day and have the subsampling respect that precise
of a request whereas the previous implementation with the `cat_valid` function
only checked the year and month of strains.