Enable preferential sampling of specific regions in strain selection
This is an alternate implementation to the original flu subsampling code [1]
that provides options to specify a region to prefer viruses from and a fraction
of viruses per month that should be sampled from that region. The original
implementation could not be copied directly into the current script because this
script is itself a reimplementation of the original method.

Closes #20.

[1] https://github.com/nextstrain/augur/blob/4ed0024/builds/flu/flu_subsampling.py#L97-L128

Co-authored-by: Richard Neher <richard.neher@unibas.ch>