Improve logic of AWS batch submission script
The existing batch.py script had initiated a separate AWS Batch job for every lineage. This resulted in each lineage asking for 5 resolutions x 2 segments = 10 CPUs, which resulted in asking for 16 CPU c5.x4large instances and thus wasted cycles. This updated script asks for 36 CPUs for a live build and 72 CPUs for a WHO build when doing a entire rebuild. This should better fill CPUs and result in cheaper billing.