Export fitness metrics via node data JSON
We want to be able to color the tree by "fitness" and "distance to future". To get these wrapped into the Auspice tree JSON we need them available as a node data JSON.

This commit revises forecast_model.py to directly output node data JSON and to extend frequencies with projected timepoints.