Include originating lab as coloring
This includes "originating lab" as coloring. I did this to make it so that when clicking on a tip it shows both submitting lab and originating lab and also so that downloading metadata gives both submitting lab and originating lab. This should provide full compliance with GISAID's Terms of Use.