Do not exclude strains last month's strains from frequency calculations
Removes configuration that prevented frequencies from being calculated for
strains sampled in the last month of the current date. This was originally
intended to apply to strain selection, too, such that both strains and
frequencies were censored to a time in the past when we thought we had
reasonable global sampling. This has since been confirmed as a Bad Idea for
strain selection, but it is also a problem for forecasts since the most recent
tips in the tree don't get frequencies calculated and then don't get forecast
into the future or assigned a weighted distance to the future.