Speed up pairwise tree titer distances by an order of magnitude
This calculation used to take ~10 minutes for a tree with 90 viruses per month
and 200 max past samples considered for each current sample. This commit
implements a more efficient approach to finding the MRCA of any two tips in the
tree and eliminates redundant walks along the path between each tip and the
MRCA. Now the same calculations complete in less than 10 seconds.