Decouple calculation of weighted distances from forecasts
Adds a new script to calculate weighted distances to the future based on a table
of tip attributes for all tips in a given tree and a table of forecasted
tips. The weighted distances are exported in the standard node data JSON format
for visualization in auspice.

This commit modifies behavior of the forecast_model.py script such that the
output table only includes the final forecast timepoint for any given list of
delta months and this table includes the amino acid sequence of the forecasted
tips. Including this sequence allows the new script to calculate pairwise amino
acid distances between tips in the tree and the projected future weighted by the
projected future frequencies.

This commit also updates the auspice config to support the new mutational load
metric and also the projected frequency output from forecasts.