Support forecasts from multiple models, keep forecast from the best model
Changes the approach to forecasting such that multiple models can be applied to
the same build and their resulting tip frequencies and forecast data frames can
be stored by model name. The user can define the models they want to use in the
config and also define the best model that should be used for the forecasts
shown in Nextstrain.

Enabling multiple forecasts and keeping their corresponding output files as part
of a normal seasonal flu build should allow us to more quickly track behavior of
multiple models and prepare figures for reports with minimal additional work.