Add support for annotating titer count per strain
Adds a script to annotate titer count per strain as either a continuous or
discrete trait (binned counts), a color-by entry in the auspice config for this
metric in the H3N2 HA 2y build, and colors for the discrete trait view. This
annotation was useful for the flu report and could be used for either HI or FRA
counts in one-off seasonal flu builds or in the private WHO builds. There is no
rule to include these annotations in any build, by default.