Remove OpenCV as a dependency for applying forecast models
OpenCV is not currently available through the Nextstrain base Docker image used
by the CLI. Importing cv2 at the top-level of the fit_model script caused live
flu builds for H3N2 HA 2y to break. This import is not necessary for making
forecasts; it is only used for fitting new models. This commit moves the cv2
import to the local context where it is used.