Stub out separate workflow to measure forecast uncertainty
Replaces the custom "select_strains.py" script with the latest augur filter
command and random subsampling up to a maximum number of samples across region,
year, and month for a fixed time period. Adds an integer wildcard for each
sampling of sequences performed. This number is used as both a random seem for
the subsampling and as an id for a given build's forecasting results.

In the end, we want to aggregate the forecasts across all samples to get a sense
of the uncertainty in our forecasts due to bias in the samples in our
analysis. This approach only begins to estimate what is a much broader
uncertainty in our sequence and titer data themselves, the phylogenetic
inference, and other models that build on these data. In the absence of a full
Bayesian model (or even a maximum likelihood model), this bootstrapping approach
should provide a better estimate of uncertainty than we have now which is
nothing.