Snakefile: include `root-sequence.json` file as target input
Renames the exported `root-sequence.json` files to the same naming
format as the final auspice JSONs and adds them as an input to the final
target rule to ensure that they are generated.

Required to have the same naming format to be found as a sidecar file
by Nextstrain.org.