Add script and rule to calculate pairwise antigenic distances
Adds a script and Snakemake rule to calculate the pairwise antigenic distance
between strains in a build. The output of the script/rule is a tidy data frame
with test strain, reference strain, raw titer distance, log2 normalized titer
distance and serum potency as well as test/reference clades, dates, and
frequencies. This table allows us to easily plot antigenic distances to
summarize antigenic advance.

Note that the rule is currently optional and must be manually requested. If we
build out visualizations based on these data, we can make these default outputs
of the workflow that get stored in a corresponding Nextstrain Group, etc.