Annotate antigenic distances with relevant columns
Annotates the source of each titer (e.g., the collaborating center) from
information in the raw titers table and adds annotations through the command
line workflow wildcards corresponding to the passage, assay, and lineage of the
data. These annotations will be helpful when we eventually aggregate all
antigenic distances within or across lineages.