calculate_epiweek: update `metadata.loc[strains]`
In our call to this script within `Snakefile_base`, the provided
metadata is for a specific segement while the provided strains list is
for all segments. This can lead to a `KeyError` when the provided
metadata is for the NA segment and the strain only has HA segement but
no NA segment.

Updating to `metadata.loc[metadata.index.isin(strains)]` ensures that
we get the inner join of the metadata and strains without worrying if
the strain exists in the metadata.