Let users define reference order in antigenic plots
Adds an argument to the script for antigenic distance plotting that accepts a
text file with a list of reference strain names in the order they should appear
in each lineage's figures. This approach makes figures more dependable and
logical in their ordering and allows us to both force-include and exclude
strains by just adding or removing them from a list.