Remove weighting of frequencies by region
Given changes in recent circulation patterns of seasonal flu, our
original region-specific weighting is no longer reasonable. For example,
although H3N2 has not been observed in China in over a year, we still
heavily weight the older strains that were observed, giving the
impression in frequency plots that its corresponding genotypes are
actively circulating when they are not. Given that we already perform
region-based subsampling to create our trees, this additional level of
weighting is not worth the noise it adds to the data.