select_strains: remove strains that do not have valid date
When parsing the metadata, completely remove a strain if it does not
have a valid date that can be converted to a numerical date. Since the
strain is directly deleted from the `tmp_meta` dictionary, we need to
loop through the list of dictionary keys instead of the iterator.

This will prevent us from running into a KeyError for `num_date` within
`flu_subsampling`.