Expand variables in subsample params for named builds
Adds logic to allow named builds (i.e., non-array-builds) to use the
same variables that array-builds support in the subsample filters. For
example, the parameter "min_date" is useful for tip frequencies and
subsampling, but named builds currently have to define the build
parameter and then repeat that information in the filter statement.
This kind of hardcoding becomes more noticeable when we'd like to have
exactly parallel subsampling logic across different sets of array
builds (as in the nextflu-private builds.)