Enable titer, fitness, distance annotations per build
Moves toggles that enable titer, glycosylation, and LBI annotations from
the top-level of the configuration file to the level of build params,
such that individual builds or sets of builds in a configuration can set
their own preferences. This will become important in the next set of
commits when we want to enable forecasts only for a specific subset of
builds.

Related to these changes, this commit also reinstates the calculation of
sequence-based distances for those lineages and segments that have
distance maps defined.