Allow users to include their own rules
As in the ncov workflow, adds the ability for users to inject their own
rules into the workflow. Although Snakemake developers seem to favor
extending an existing workflow with modules [1], this custom rules
approach works well enough when we expect users (e.g., ourselves) to
have cloned the workflow repo already anyway. We might decide this
approach is a bad idea in the future, but for now it is very useful.

[1] https://snakemake.readthedocs.io/en/stable/snakefiles/modularization.html#modules