Add "is_reference" column to metadata
Adds a boolean column "is_reference" to the metadata used for each
lineage, based on the list of reference strains in the config for that
lineage. This additional column allows us to define subsampling schemes
that select reference strains over a different time period than the rest
of the data in an analysis, recreating the effect of the
`--max-include-range` argument from `select_strains.py` in the original
workflow.