Annotate strains with titer data in metadata
Enables subsampling in builds that create a measurements panel based on
whether strains have titer measurements or not. Two new optional rules
annotate the full metadata with a boolean column `is_titer_strain` using
the titers requested for each build. We can then use this column to
define a subsampling scheme that selects for all titer strains in a
given date range, as shown by the updated nextflu-private configuration
here. This approach allows us to get around an issue where subsampling
with prioritization of strain by titer counts does not guarantee that
all titer strains will get included in a build. We also do not want to
force-include all titer strains, since many strains date back to times
prior to the rest of the strains in a given build. The approach here
allows us to effectively force-include titer strains for a given date
range. This approach should have no effect on builds that do not use
a measurement panel.