Finalize forecast aggregation
Updates the logic to convert forecast frequencies from Auspice JSONs to
TSVs to use a script originally developed for the VCM reports that does
a better job of annotating forecasted frequencies by all possible
clades. Also updates the workflow logic to aggregate the new forecast
tables by model for Zoltar and across all models. Updates Zoltar
preparation script to read a single aggregated table of forecasts for a
single model and find active clades to generate outputs for.

This commit also adds a script used to generate plots of the final
forecasts, although it does not yet add a rule in the workflow to
generate these plots automatically.