Add reference outgroup for 2y/6m builds
Add reference outgroup for 2y and 6m builds to avoid incorrect rooting
of trees. The reference is pruned from the tree after the refine step
based on James' code snippet¹.

The references added are the references used by Nextclade, but with the
strain names changed to match the strain name within fauna.

¹ https://github.com/nextstrain/augur/issues/340#issuecomment-545184212