Fix inputs for haplotype annotation
We can't define the translations with internal nodes as an input to any
rules like the haplotypes rule, because no rule produces these files as
explicit outputs. Although we should fix this, the short-term fix is to
ask for the `translations.done` flag file and then refer to the
alignment FASTA from the params block of any rule that needs this file.