Add public version of nextflu-private builds
These builds are the same as nextflu-private builds in that they only
produce recent trees for the three major lineages, but they do not build
egg-passaged trees for all lineages or HI-based trees for H3N2 and they
do not include the measurements panel or antigenic distance plots. These
changes allow the resulting builds to be shared publicly.