Convert CI build to use titer collections
Drops whitespace from the rows in example titer data, adds a header to
example titers, and updates the CI configs to use these data for the
builds. Since we provide a specific path to the data we want to use for
titers, we no longer need a rule to "prepare titers".