Make titer attribute prefix optional
Revert to the original default behavior for the titer model subcommands
which is to not use the attribute prefix argument. This implementation
adds a new `prefix` attribute to each titer collection. When the
attribute is not defined, the workflow does not provide an attribute
prefix argument to the titers subcommands. Previous to this commit, the
workflow would use the titer collection name as the attribute prefix
without any option to opt out. This new logic allows all existing builds
to define a name and data path without also defining an empty prefix
string to opt out. Builds that use multiple titer collections and need
the prefix must manually add that prefix to each titer collection.

Since this change maintains the original node data attribute names for
titer model outputs (e.g., `cTiter`), we do not need to modify the
Auspice config JSONs for H3N2 to refer to a hardcoded alternate prefix
for these attributes.