Rename public builds to match current format
Adds an `auspice_name` build parameter to each public build with the
format expected by the current seasonal flu
site (`flu_seasonal_{lineage}_{segment}_{resolution}`) and adds custom
rules to handle the renaming for each type of Auspice JSON from the
`{build_name}_{segment}` format. An annoying technical detail here is
that we need to add a `resolution` block to each array build with a
single entry, so we can reuse the build params across all resolutions
and get the proper resolution value in the `auspice_name`. This
implementation is a little more complex than just bulk copying all the
output files to rename them, but it plays nicely with Snakemake's
expectations of input/output files and only needs to be written
once (now) and then hopefully very little editing needs to happen to it
in the future.