Prune reference strain prior to time tree
Adds and modifies a script from master branch that prunes the reference
strain from a given tree. The script now accepts a FASTA file for the
reference used for alignment, so the name of the reference can be pulled
in programatically. The reference pruning occurs prior to augur refine,
so the reference does not affect time tree calculations.