Clean up H3N2 NA reference name
Changes name of the reference sequence to an unambiguous string that
remains the same before and after `augur tree`. The previous string had
a ":" that got converted to an "_" by the `tree` command preventing us
from matching the reference sequence by name between the alignment FASTA
and the Newick file.