Add config, rules, and scripts to support private.nextflu.org
Copies most rules from Snakefile_WHO in the original workflow which
creates the JSON files used by private.nextflu.org and updates these
rules to work in the context of the refactored workflow. Includes
Auspice config JSONs from the master branch and two scripts that we need
for calculating the "entropy" side-car file and the scores node data
JSON. Adds a build configuration file for the nextflu site that matches
the builds produced by the "batch.py" script.

Note that rules required to create the "frequencies" panel have been
commented out for now. We should be able to add these next.