Allow regions to be superset of given frequencies
Allows the list of regions to be a superset of the list of frequencies
available per region, to handle the case when a given region has no data
for a build and therefore has no frequencies. The new logic checks for
the presence of each given region name (e.g., "Japan Korea") in a given
frequencies JSON filename to find a list of available regions.