Estimate global frequencies from non-empty region translations
Fixes a data-dependent bug when the workflow tries to estimate
region-specific frequencies for a region that has no aligned amino acid
sequences. This fix includes two important steps:

1. augur filter run on amino acid sequence alignments to select data for
a specific region is allowed to have empty output (printing only a
warning, so we can track these events in logs), so the workflow can
create an empty FASTA file for regions without data. These empty files
represent that state of missing data. This implementation uses a new
argument for augur filter provided in version 21.1.0. This commit
updates the Conda environment accordingly.

2. The workflow conditionally builds a list of regional mutation
frequencies to estimate for each build based on which amino acid
sequence files are not empty. Because there can be multiple "genes" per
segment, the workflow requires each region to have non-empty amino acid
sequences for all genes in a segment. This implementation relies on
converting the augur filter rule to a "checkpoint", so Snakemake knows
to update the DAG after that rule runs. The aggregation of which region
frequencies to estimate happens by calling an input function to the
final global frequencies rule that checks the content of the filtered
FASTA files and builds the list of region frequency JSONs to create.

This last step works because the global frequency estimation script
accepts a list of all possible regions and a list of available frequency
JSONs.