Annotate year/month for easier date filtering
Adds a `year_month` annotation through the existing `epiweeks`
annotation that provides a coarser grain filtering by date. This kind of
filtering is most useful in narrative writing, where we want to easily
update a prior filter query string to reflect the latest data. Adding a
new month to a list of year/months is much easier than adding all of the
epiweeks.