Correct and document Vic HA gene coordinates
Corrects the HA gene coordinates for B/Vic such that HA1 and HA2
translations have the expected lengths (347 and 223, respectively) where
previously the HA1 sequence was 1 amino acid longer at the end and the
HA2 sequence was 1 shorter at the beginning. Documents the source of the
new coordinates with the PDB record and corresponding publication.