plot_titer_matrices: remove reference to `select_strains`
We've removed the `select_strain` script in 1cedac10b163c90a5128d3774e20feaf5fd48d6a,
so remove the reference here and the unused imports from the script.

Also removes the commented code that relies on the metadata created
from `parse_metadata`. If needed in the future, we can bring it back
through Git history.