Use modern Snakemake profile for SLURM jobs
Updates the Snakemake profile for the Hutch's SLURM environment to use the
modern Snakemake interface for requesting cluster resources and submitting
jobs that can be easily cancelled. Adds runtime resource constraints to jobs
that need them and which were previously defined in the "cluster.json" file
in the Hutch's profile directory. These changes should have no effect on
execution of the workflow in other environments except to provide those
constraints per rule in a way that can be reused in other environments.