Update README for refactored workflow
Refactoring the workflow changed organization of the repo and the
Snakemake arguments used to run the builds. This commit updates the
docs in the README to match the refactored workflow and uses the
CI build as an example workflow for external users.

Motivated by https://github.com/nextstrain/seasonal-flu/issues/112.