nextstrain-public/all_deploy: remove log directive
We were running into the Snakemake issue where the log directive in
the default rule causes it to skip tasks.¹ This issue has been fixed
since Snakemake 7.25.1, but we are still pinned to 7.24.1 in
nextstrain/docker-base.²

Just removing the log directive here is the simplest fix for this
particular issue. We'll still be able to see the upload logs in
the AWS Batch logs.

¹ https://github.com/snakemake/snakemake/issues/2142
² https://github.com/nextstrain/docker-base/blob/95cea4f44a803ce27992706c87a7237492dbb083/Dockerfile#L277