Fix reference to build params in clade rules
Replaces use of a variable, `build_dir`, with the hardcoded key,
"builds", that we consistently use to store build parameters in our
config YAMLs. The references changed here only worked because
`build_dir` was set to "builds", too, but if we ever changed the build
directory, the workflow would break.