Keep reference strains in tree even if they are outliers
Adds an argument to the script for flagging outliers that accepts a file
with a list of strains to keep in the output tree even if they are
outliers and updates the core workflow logic to use this argument
whenever builds define a list of strains to force-include. This change
fixes a bug when older titer reference strains (e.g.,
A/Wisconsin/588/2019 for H1N1pdm) get flagged as outliers in the tree
and dropping them causes downstream titer analyses to fail. Our choice
to force-include these strains in the build should override the
automated outlier detection.