Update builds to use `--env` options
Better to be explicit about which envvars are being passed to runtimes
rather than rely on the automatic variables passed through.

The upload workflow requires the rethinkdb secrets to access Fauna
and the AWS secrets to upload files to S3.

The public and private builds just need the AWS secrets to launch jobs
on AWS Batch and to download files from S3.

These are all GitHub Action secrets that are being inherited by the
reusable workflow and available as envvars in the GitHub Action job.