Add workflow for private.nextflu.org deployments
This basic workflow is able to build the empty site and deploy to
staging-private-nextflu.netlify.app, which is still password protected.

The private.nextflu.org site is still built using the old Auspice
version in https://github.com/blab/nextflu and deployed to Netlify.
This requires an older version of Ruby (2.3.7) and dependencies that
others have had trouble installing on the new M1 macs. Instead of
depending on my old mac to deploy the private builds, create to a GitHub
Action workflow  to deploy the site that anyone on the Nextstrain team
can run.

I've opted for this route instead of trying to update the dependencies
for the old Auspice because the ultimate goal is to sunset the private
site. This seemed like less work than fiddling with dependencies.