Add script to generate color scales
Adds a script to generate an Auspice config JSON with a custom color
scale for a specific coloring field in an existing JSON and a list of
user-provided values that should receive colors. Uses an existing color
schemes file (`config/color_schemes.tsv`) to assign colors to the
requested values after sorting those values alphabetically.