Use custom script to convert XLS to CSV
Since xlrd is already available in the Nextstrain Docker image, use this
lower-level interface to convert the GISAID metadata from XLS to CSV
with a small custom script. The Nextstrain Conda environment does not
install xlrd, but this commit adds "xlrd" to the workflow-specific
environment, allowing Conda users to run Snakemake with the
`--use-conda` flag inside the Nextstrain Conda shell.

This approach avoids installing an extra dependency like csvkit in the
Nextstrain Docker and Conda environments and avoids complexity
associated with pandas optional dependencies.