Use tabulate's float formatting for tables
Instead of rounding decimal values prior creation of the Markdown table
output, use tabulate's own `floatfmt` argument to control that display
detail. Since we used to round current frequencies to 2 decimals and
then filter haplotypes with frequency of zero, we now filter for
haplotypes with a current frequency of at least 1% which produces the
same effect but in a clearer way.