Add build config for Yamagata
Adds a custom build config for Yamagata public builds using hardcoded
dates, since relative dates do not work for an extinct lineage. This
config allows us to regenerate Yamagata builds in the future, though, to
change annotations or otherwise update the builds to match the other
seasonal influenza lineages.

In this case, we need to re-run the Yamagata builds to use the latest
gene map and fix the broken entropy panel in the public builds first
reported on the discussion
forum (https://discussion.nextstrain.org/t/yamagata-flub-ha1-amino-acid-in-the-genotype-or-color-by-selections/1553/2).

Related to #57