Add Welsh escape model for forecasts
Adds a simple forecast model based on Welsh escape scores using a
coefficient of 1.0 that assumes escape scores correspond directly to
fitness. This is nearly the same coefficient as LBI in the LBI and
mutational load model. We use this approach for now as a way to account
for the current frequencies of clades along with the escape scores per
clade. In the future, we will want to learn historical coefficients for
these data, ideally from a dataset spanning a longer time period.