Download clade definitions from GitHub
Remove local copies of clade and subclade definitions for H1N1pdm, H3N2,
and Vic and add configuration and rules to download these files from the
corresponding GitHub repositories as needed by each workflow. This
approach allows us to maintain local copies of clades and subclades for
Yam and other exceptions to the standard maintained clades. Since we
only need to maintain the two "clade systems" of "clade" and "subclade"
for the main phylogenetic workflow, this commit defines a top-level
mapping of clade file URLs to lineages and segments. This is in
contrast to the build-specific clade mappings used by the Nextclade
workflow [1].

[1] https://github.com/nextstrain/seasonal-flu/blob/d439f897f2819961dee431f92ea4e6f45343688a/nextclade/config/config_dict.yaml#L8-L16