Remove H3N2 vaccine strain from outliers
I noticed today when looking for A/HongKong/45/2019 in the 12y tree that
I accidentally added this strain to the outliers file back in June
2023 (commit 2091f667) when I included all outliers detected by the
"local outlier" script. This was before the local outlier detection had
the concept of force-including reference strains. This commit undoes
that mistake. A subsequent commit will prevent reference strains from
getting excluded from builds even if this kind of mistake happens again.