Simplify Nextclade build config
Remove unnecessary configuration parameters from the Nextclade build
config and update the workflow to allow these parameters to be missing.
Since Snakemake evaluates the Python code in each rule's inputs,
outputs, and params, rules that we don't plan to run in the workflow can
produce key errors when their config parameters are not defined.