Simplify Nextclade dataset logic
Simplifies the logic to get Nextclade datasets by following the same
pattern as the flu_frequencies workflow [1] where we grab the default
dataset for a given lineage and segment instead of specifying a
reference name. The "broad" and more recent references for H3N2 HA, for
example, are not too different from each other, but the Nextclade
annotations for the former are far more verbose than for the latter. We
also want the files produced by this workflow to plug directly into the
flu_frequencies workflow logic, so it is best to use the same approach
here.

[1] https://github.com/nextstrain/flu_frequencies/blob/6e4298fac3361f4a6751d85bcb963064dbb9eee1/Snakefile#L95