Make example build a starting point for new users
Converts the "example" build configuration from one that depends on
access to our internal database (fauna) to one that depends only on the
example data provided by this repository. New users can run this
workflow as a quickstart before moving on to the GISAID-based
quickstart, since this workflow doesn't require users to curate any
data.

The CI workflow is not a good quickstart for users, since it analyzes
titer data that most people will not have at the start of their
analysis.