Add isolate_id field to metadata
Adds the field containing GISAID isolate id (values prefixed by
"EPI_ISL_") to the list of fields to include in the download from fauna
in the "upload" configuration file. When we run the upload workflow with
this change, the parsed metadata files on S3 should include the new
column without any other changes needed by other workflows that download
the data from S3.

This isolate id field is important in that it represents a single viral
isolate that has one or more segment accessions for genes like HA, NA,
etc.