Count all recent sequences with Nextclade TSVs
Modifies the script that makes a Markdown table of recent sequence
counts per clade to use a recency JSON constructed from all metadata per
lineage (instead of metadata for a specific tree) and the Nextclade
annotations on S3. This change allows us to report the clade counts for
all recently submitted sequences instead of limiting our report to
sequences that appear in a given tree.

Related to #130