Add representative samples from early H1 HA clades
Extends the min date for the "broad" H1N1pdm HA dataset from 2014 to
2009 and adds manually curated representative strain names for early
clades 2, 3, 4, 7, and 8 to the "force-include" list for the Nextclade
workflow. These changes allow the broad Nextclade dataset to represent
most early clades (except clade 1 [1]) such that early sequences can be
properly assigned to those clades.

This approach of forcing inclusion of representative strains works
around the workflow's filter of QC=bad sequences where the QC is based
on the more recent Nextclade dataset. Since that dataset lacks early
clades, early sequences from those clades map to the newer tree with too
many private mutations and get flagged with bad QC. A better approach
could be to run Nextclade with the "broad" dataset for each lineage, to
minimize the number of false positive bad QC labels, but that is for a
future discussion/PR.

[1] https://github.com/influenza-clade-nomenclature/seasonal_A-H1N1pdm_HA/issues/2