Add representative samples from clade 1
Now that clade 1 has an official entry in the clade nomenclature, we can
add representative samples for this clade to be force-included to the
Nextclade tree for the broad H1N1pdm HA dataset. Including these samples
produces a tree where the MRCA is clade 1 and all other clades descend
from this clade.