Estimate frequencies for all metadata records
Adds a script and rule to estimate "tip" frequencies JSON from metadata
alone. This simple functionality isn't provided directly through `augur
frequencies`, so this commit adds a script that replicates some of the
internal logic of that Augur script to get a tip frequencies JSON with
the KDE-based method. Since KDE frequency estimates only require a list
of dates, we can estimate frequencies for each sequence in the metadata
and use those estimates in a subsequent rule to estimate frequencies of
derived haplotypes.

In this commit, I chose to limit the frequency estimation period to a
max date of 4 weeks prior to the current run date and a min date 16
weeks prior. These frequencies will only be used initially to compare
the most recent value to the timepoint just previous to calculate a
delta frequency.