Use same haplotype style for tree and tables
Updates the script that annotates derived haplotypes for nodes in the
tree to use the same style as the haplotypes table with hyphen-delimited
mutations (which work as values in URL parameters unlike comma-delimited
lists) and with the ancestral allele included for each mutation. These
changes should allow us to link from the haplotype tables to the tree
view for the same haplotypes.