Add t-SNE embeddings and clusters to workflow
Adds logic and scripts to the core workflow to create a t-SNE embedding
per build with all segments and create HDBSCAN clusters from the
embeddings. Updates the Auspice config JSONs the public and private
builds to include the new t-SNE coordinate fields and the t-SNE cluster
field.

Although we select the same strains for each gene segment, not all
segment sequences can be aligned. As a result, different sets of strains
can appear in the final Nextclade alignment for each gene segment. Since
the t-SNE embedding requires the same strains to be present for all of
its inputs and in the same order, we need to find the list of strains
shared across all segments, extract only those strain sequences from
each segment alignment, and sort the alignments prior to calculating
pairwise distances and the embedding. We could perform the necessary set
operations on multiple segment alignments with shell commands only, but
this implementation was not as readable as the simple Python script
included here.

Closes #174