Only create embeddings when opting in
Make the introduction of t-SNE embeddings more backward-compatible by
making them an opt-in feature of the workflow like titer models, LBI,
etc. Enables embeddings for CI (to catch errors) and for public and
private builds for nextstrain.org. We may further decide that the main
value of these embeddings is in the 2y builds and modify the
nextstrain-public config to only run embeddings for those builds.