Rename metadata column from `isolate_id` to `gisaid_epi_isl`
The fields to export in acknowledgements is hardcoded in Auspice¹ so the
metadata field needs to be named `gisaid_epi_isl` to be included in
the acknowledgements for the seasonal flu builds. Renaming the column
in our upload config to avoid having to rename it as part of the
phylogenetic workflow.

¹ <https://github.com/nextstrain/auspice/blob/dae78e74d5e06303d901387adf9fe7550a7d4aa9/src/components/download/helperFunctions.js#L326>