Support derived haplotypes at a subset of sites
Allows users to pass a distance map JSON to the derived haplotype
calculation script to specify a subset of genes and positions to
consider as part of each derived haplotype. The benefit of this
approach is that we can define derived haplotypes for epitope sites or
RBD sites, for example.