Run Nextclade on all 8 genes for H1 and H3
Adds remaining gene segments to the default list of segments to process
with Nextclade. Since B/Vic only has HA and NA Nextclades datasets right
now, this commit modifies the logic of the workflow to look for a
build-specific list of segments and defines that list for the Vic build.