Add proposed subclade coloring to non-2y builds
Adds a coloring for the proposed subclade annotation in the Auspice
config JSONs used by all non-2y public builds. This coloring already
appears in the builds as "proposed_subclade" at the end of the colorings
list , but this commit makes "Proposed subclade" appear right after
"Subclade" in the coloring list.