Limit concurrent fauna connections to 2
The latest upload workflow¹ seemed to overwhelm the rethinkdb server
so limiting concurrent fauna connections to see if it would prevent
the OOM error.

Use of Snakemake's `workflow.global_resources` inspired by the
ncov use to limit concurrent deploy jobs.² Doing this instead of
limiting the overall number of Snakemake jobs with `-j` because there
are other rules in the workflow that can run without connections to
fauna.

¹ <https://github.com/nextstrain/seasonal-flu/actions/runs/12715154181/job/35446799658>
² <https://github.com/nextstrain/ncov/blob/20f5fc3c7032f4575a99745cee3238ecbeebb6e0/workflow/snakemake_rules/export_for_nextstrain.smk#L340-L362>