Force-include old titer reference strains
Adds a new column to the metadata indicating whether a strain is a titer
reference strain or not and updates the subsampling logic for private
builds to force-include titer reference strains as far back in time as
other reference strains. This change fixes an issue where the 2-year
limit on titer strains would cause reference strains to be dropped from
builds and cause their titer measurements to also be dropped even if
those measurements were against more recent test viruses. Historically,
we manually maintained a list of reference strains per subtype to make
sure these kinds of older titer references were force-included.
Maintaining this list isn't sustainable or necessary, since we can
automate the process like we do in this commit.